% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date2fy.R
\name{date2fy}
\alias{date2fy}
\title{Convert date to fiscal year}
\usage{
date2fy(date, fye = FALSE, separator = "/")
}
\arguments{
\item{date}{A Date or POSIXct vector}

\item{fye}{A logical. If TRUE returns fiscal year ending}

\item{separator}{A character. Symbol to separate fiscal year starting and ending (forward slash by default)}
}
\value{
fiscal year (character by default. Numeric if fye is TRUE).
}
\description{
Convert a date to fiscal or financial year. UK fiscal year runs from 1 April to 31 March.
}
\examples{
date2fy(as.Date("2024-05-07"))
}
\author{
Simon Maxwell
}
