% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date2py.R
\name{date2py}
\alias{date2py}
\title{Convert date to planting year}
\usage{
date2py(date, pye = FALSE, separator = "/")
}
\arguments{
\item{date}{A Date or POSIXct vector}

\item{pye}{A logical. If TRUE returns planting year ending}

\item{separator}{A character. Symbol to separate planting year starting and ending (forward slash by default)}
}
\value{
Planting year (character by default. Numeric if pye is TRUE).
}
\description{
Convert a date to planting year or season. UK tree planting season runs from 1 October to 30 September.
}
\examples{
date2py(as.Date("2024-05-07"))
}
\author{
Simon Maxwell
}
