% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_alphabetical.R
\name{is_alphabetical}
\alias{is_alphabetical}
\title{Check if letters in word are in alphabetical order}
\usage{
is_alphabetical(word)
}
\arguments{
\item{word}{A word.}
}
\value{
A logical.
}
\description{
Consecutive identical letters are treated as maintaining alphabetical order.
}
\examples{
is_alphabetical("aegilops")
}
\author{
Simon Maxwell
}
