% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxwllibrary-package.R
\docType{package}
\name{mxwllibrary-package}
\alias{mxwllibrary}
\alias{mxwllibrary-package}
\title{mxwllibrary: Personal R package}
\description{
Collection of random function I have made.
}
\author{
\strong{Maintainer}: Simon Maxwell \email{simonjosephmaxwell@gmail.com}

}
\keyword{internal}
