% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_nwd_statement.R
\name{print_nwd_statement}
\alias{print_nwd_statement}
\title{Print non-working day statement for work email signature}
\usage{
print_nwd_statement(block_start, n)
}
\arguments{
\item{block_start}{Date. The date the two-week block started (must always be a Monday).}

\item{n}{numeric. The number of Fridays to print.}
}
\value{
character
}
\description{
I work a 9-day fortnight with every second Friday off.
}
\examples{
print_nwd_statement(as.Date("2024-06-17"), 3)
}
\author{
Simon Maxwell
}
